// ClientApp/generated/MailReqModel.ts
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */
import { PreparePdfresultModel } from './PreparePdfresultModel';
//Generated from class components.printApp.MailReqModel
export interface MailReqModel {
  mailtoList: string[];
  ccList: string[];
  message: string;
  subject: string;
  attachmentName: string;
  pdfJobResult?: PreparePdfresultModel;
}